Use \Closure instead of callable
callable is a. slightly slower and b. subject to scope in PHP, as I
learned from a commit message of TimWolla.

Being kind and helpful, both TimWolla and Edorian helped explain and
benchmark this a little to get some intuition.

This keeps callable only in functions like `fmap` and `c` that in
general are for DX but changes the interfaces to use \Closure only.