WIP on assert functions
There are multiple issues here.

The tests seem to be wrong, called with wrong functions.

Then there are no tests for the assertions.

Then there is an issue with equality as it depends on the
implementation of the functor. For example, it works on Maybe but not
on the others, due to internal structure.