Default to run the passed function in Eq1::liftEq
Not quite sure if this is good, however if you were to lift equality
function on type that is represented without an object AND there is no
registered instance you were getting a type error. It is trivial
though to be able to do something like:

eq1($one, $two)

And assume that the `==` will be called (equals()).